shf_tools/sgf2shf/requirements.txt
PyQt6>=6.4.0
sgf==0.5
sgfmill>=1.1.1
opencc-python-reimplemented>=0.1.7 